Raw font: make sym2raw/raw2sym standalone functions
This replaces the previous `applyFontFilter` and `unapplyFontFilter` functions
and call sites have been adjusted accordingly.
